Consolidate lesson report pages.
